Revert dictation HUD side-zone upright orientation
The quarter-turn belongs to the meeting recording pill; on the dictation
HUD it was more acrobatics than utility (Wispr doesn't even let you move
theirs). Reverts the four commits that built it:

- dictation HUD: lay flat before text states paint; glide the edge
  clamp; reveal gate
- dictation HUD: shorter upright pill, kept on-screen, pre-rotated at
  show
- dictation HUD: apply the upright square synchronously so the turn
  isn't clipped
- dictation HUD: stand upright in side zones while listening

The dictation pill keeps everything that matched it to the meeting pill:
the frosted vibrancy surface (back via window_vibrancy now that nothing
needs a pinned frost), the darker tint, window-sized-to-pill morphs, the
native alpha exit and shake. The meeting HUD keeps its upright turn,
with its native machinery folded back into meeting_hud.rs (hud_native.rs
removed).
