+++ src-tauri/src/lib.rs
@@ -5,7 +5,6 @@ pub mod db;
 pub mod dictation;
 pub mod domain;
 pub mod hermes_bridge;
-pub mod hud_native;
 pub mod mascot;
 pub mod meeting_detection;
 pub mod meeting_hud;
@@ -125,8 +124,6 @@ pub fn run() {
             dictation::dictation_hud_set_size,
             dictation::dictation_hud_set_alpha,
             dictation::dictation_hud_shake,
-            dictation::dictation_hud_flatten,
-            dictation::dictation_hud_apply_zone,
             dictation::dictation_hotkey_status,
             dictation::latest_dictation_event,
             mascot::mascot_show,
