+++ src-tauri/src/meeting_hud.rs
@@ -28,12 +28,14 @@
 
 use crate::audio::capture;
 use crate::domain::types::{RecordingState, RecordingStatusDto};
-use crate::hud_native::{self, zone_for, Zone, TURN_SECS};
 use std::{sync::Mutex, thread, time::Duration};
 use tauri::{
     AppHandle, Emitter, LogicalSize, Manager, PhysicalPosition, PhysicalSize, State, WebviewWindow,
     WindowEvent,
 };
+
+#[cfg(target_os = "macos")]
+use objc2::runtime::{AnyClass, AnyObject};
 
 const WINDOW_LABEL: &str = "meeting-hud";
 
@@ -53,9 +55,27 @@ const VERTICAL_PILL_LENGTH: f64 = 62.0;
 /// gutters above/below (or beside) the pill are part of the window.
 const WINDOW_SIZE: LogicalSize<f64> = LogicalSize::new(76.0, 76.0);
 
+/// How long the quarter turn takes. The easing matches the app's `--ease-out`
+/// token (cubic-bezier(0.22, 1, 0.36, 1)) so the HUD moves like the rest of
+/// the UI even though this animation runs in Core Animation, not CSS.
+const TURN_SECS: f64 = 0.32;
+
 /// How many supervisor ticks the window must hold still before an orientation
 /// flip is applied — resizing mid-native-drag is what we're avoiding.
 const SETTLE_TICKS: u32 = 4;
+
+#[derive(Debug, Clone, Copy, PartialEq, Eq)]
+pub enum Zone {
+    Left,
+    Center,
+    Right,
+}
+
+impl Zone {
+    fn is_vertical(self) -> bool {
+        matches!(self, Zone::Left | Zone::Right)
+    }
+}
 
 pub struct MeetingHudPosition {
     /// Drag position remembered between appearances in the same process. Like
@@ -197,6 +217,29 @@ fn supervise(app: &AppHandle, tracker: &mut ZoneTracker) -> Duration {
     ACTIVE_TICK
 }
 
+/// Which third of the monitor's work area the pill center sits in.
+fn zone_for(hud: &WebviewWindow) -> Option<Zone> {
+    let position = hud.outer_position().ok()?;
+    let size = hud.outer_size().ok()?;
+    let center_x = position.x + size.width as i32 / 2;
+    let center_y = position.y + size.height as i32 / 2;
+    let monitor = hud
+        .monitor_from_point(center_x as f64, center_y as f64)
+        .ok()
+        .flatten()
+        .or_else(|| hud.current_monitor().ok().flatten())?;
+    let work = monitor.work_area();
+    let third = work.size.width as i32 / 3;
+    let offset = center_x - work.position.x;
+    Some(if offset < third {
+        Zone::Left
+    } else if offset > 2 * third {
+        Zone::Right
+    } else {
+        Zone::Center
+    })
+}
+
 /// Watch the drag position; once the window has held still for a few ticks
 /// AND the mouse button is up, flip orientation if it crossed into a different
 /// third. Both guards exist so we never resize mid-drag — the resize would
@@ -212,7 +255,7 @@ fn track_zone(hud: &WebviewWindow, state: &MeetingHudState, tracker: &mut ZoneTr
         tracker.stable_ticks = 0;
         return;
     }
-    if tracker.stable_ticks >= SETTLE_TICKS && !hud_native::left_mouse_button_down() {
+    if tracker.stable_ticks >= SETTLE_TICKS && !left_mouse_button_down() {
         // Apply once per rest: bump past the threshold so we don't re-run
         // zone math every tick while parked.
         if tracker.stable_ticks == SETTLE_TICKS || tracker.pending_release {
@@ -223,6 +266,24 @@ fn track_zone(hud: &WebviewWindow, state: &MeetingHudState, tracker: &mut ZoneTr
         // Settled but still held — remember to flip on release.
         tracker.pending_release = true;
     }
+}
+
+/// Whether the left mouse button is currently held, via the window server's
+/// combined session state — works regardless of which window has the cursor.
+#[cfg(target_os = "macos")]
+fn left_mouse_button_down() -> bool {
+    #[link(name = "CoreGraphics", kind = "framework")]
+    extern "C" {
+        // CGEventSourceButtonState(kCGEventSourceStateCombinedSessionState = 0,
+        //                          kCGMouseButtonLeft = 0)
+        fn CGEventSourceButtonState(state_id: i32, button: u32) -> bool;
+    }
+    unsafe { CGEventSourceButtonState(0, 0) }
+}
+
+#[cfg(not(target_os = "macos"))]
+fn left_mouse_button_down() -> bool {
+    false
 }
 
 /// Recompute the zone and, if it changed, turn the pill to the new orientation.
@@ -257,33 +318,9 @@ fn apply_zone_now(hud: &WebviewWindow, state: &MeetingHudState) {
 /// shadow is recomputed once the turn lands.
 #[cfg(target_os = "macos")]
 fn set_orientation(hud: &WebviewWindow, vertical: bool, animate: bool) {
-    use objc2_foundation::{NSPoint, NSRect, NSSize};
-
     let window = hud.clone();
     let _ = hud.run_on_main_thread(move || unsafe {
-        let angle = if vertical {
-            // CA's +z spins counterclockwise (y-up); -90° swings the pill's
-            // left end — the record dot — to the top.
-            -std::f64::consts::FRAC_PI_2
-        } else {
-            0.0
-        };
-        hud_native::rotate_content(&window, angle, animate);
-        // The upright pill runs shorter (fewer bars); keep the frost hugging
-        // it. Same duration and curve as the turn, on the same CA clock.
-        let length = if vertical {
-            VERTICAL_PILL_LENGTH
-        } else {
-            PILL_SIZE.width
-        };
-        let frost_frame = NSRect::new(
-            NSPoint::new(
-                (WINDOW_SIZE.width - length) / 2.0,
-                (WINDOW_SIZE.height - PILL_SIZE.height) / 2.0,
-            ),
-            NSSize::new(length, PILL_SIZE.height),
-        );
-        hud_native::set_frost_frame(&window, frost_frame, false, animate);
+        rotate_content(&window, vertical, animate);
     });
     if animate {
         // The shadow shape is derived from the rendered content; refresh it
@@ -293,7 +330,7 @@ fn set_orientation(hud: &WebviewWindow, vertical: bool, animate: bool) {
             thread::sleep(Duration::from_secs_f64(TURN_SECS + 0.06));
             let handle = window.clone();
             let _ = window.run_on_main_thread(move || unsafe {
-                hud_native::invalidate_shadow(&handle);
+                invalidate_shadow(&handle);
             });
         });
     }
@@ -318,20 +355,11 @@ fn configure_window(app: &AppHandle) -> Result<(), String> {
 
         #[cfg(target_os = "macos")]
         {
-            use objc2_foundation::{NSPoint, NSRect, NSSize};
-
             make_nonactivating(&hud);
             // Real behind-window blur, sized to the pill (not the window — the
-            // square window's gutters stay fully transparent). The 10px radius
-            // must match the pill's CSS border-radius (--r-lg).
-            let pill_rect = NSRect::new(
-                NSPoint::new(
-                    (WINDOW_SIZE.width - PILL_SIZE.width) / 2.0,
-                    (WINDOW_SIZE.height - PILL_SIZE.height) / 2.0,
-                ),
-                NSSize::new(PILL_SIZE.width, PILL_SIZE.height),
-            );
-            unsafe { hud_native::install_frost(&hud, Some(pill_rect), 10.0) };
+            // square window's gutters stay fully transparent). Done by hand
+            // rather than via window_vibrancy, which always fills the window.
+            unsafe { install_frost(&hud) };
         }
 
         let app_for_events = app.clone();
@@ -416,13 +444,248 @@ fn position_is_visible(
     })
 }
 
+/// The NSWindow's contentView as a raw object, or null.
+#[cfg(target_os = "macos")]
+unsafe fn content_view(hud: &WebviewWindow) -> (*mut AnyObject, *mut AnyObject) {
+    use objc2::msg_send;
+
+    let Ok(handle) = hud.ns_window() else {
+        return (std::ptr::null_mut(), std::ptr::null_mut());
+    };
+    if handle.is_null() {
+        return (std::ptr::null_mut(), std::ptr::null_mut());
+    }
+    let window = handle as *mut AnyObject;
+    let content: *mut AnyObject = msg_send![window, contentView];
+    (window, content)
+}
+
+/// macOS: add an NSVisualEffectView behind the webview, framed to the pill
+/// rather than the window, so the square window's gutters stay fully
+/// transparent. Hand-rolled instead of `window_vibrancy::apply_vibrancy`
+/// because that always fills the window.
+#[cfg(target_os = "macos")]
+unsafe fn install_frost(hud: &WebviewWindow) {
+    use objc2::msg_send;
+    use objc2_foundation::{NSPoint, NSRect, NSSize};
+
+    /// Must match the pill's CSS border-radius (--r-lg, 10px) so tint and
+    /// frost trace the same curve.
+    const FROST_RADIUS: f64 = 10.0;
+
+    let (_, content) = content_view(hud);
+    if content.is_null() {
+        return;
+    }
+    // The contentView hosts the orientation turn (`rotate_content`); give it a
+    // layer up front so the first turn doesn't switch rendering paths mid-use.
+    let _: () = msg_send![content, setWantsLayer: true];
+
+    let Some(effect_class) = AnyClass::get(c"NSVisualEffectView") else {
+        return;
+    };
+    let frame = NSRect::new(
+        NSPoint::new(
+            (WINDOW_SIZE.width - PILL_SIZE.width) / 2.0,
+            (WINDOW_SIZE.height - PILL_SIZE.height) / 2.0,
+        ),
+        NSSize::new(PILL_SIZE.width, PILL_SIZE.height),
+    );
+    let frost: *mut AnyObject = msg_send![effect_class, alloc];
+    let frost: *mut AnyObject = msg_send![frost, initWithFrame: frame];
+    if frost.is_null() {
+        return;
+    }
+    // HUDWindow material, behind-window blending, always active — the same
+    // look the dictation HUD gets from window_vibrancy.
+    let _: () = msg_send![frost, setMaterial: 13isize];
+    let _: () = msg_send![frost, setBlendingMode: 0isize];
+    let _: () = msg_send![frost, setState: 1isize];
+    let _: () = msg_send![frost, setWantsLayer: true];
+    let layer: *mut AnyObject = msg_send![frost, layer];
+    if !layer.is_null() {
+        let _: () = msg_send![layer, setCornerRadius: FROST_RADIUS];
+        let _: () = msg_send![layer, setMasksToBounds: true];
+    }
+    // Below the webview (NSWindowBelow = -1) so the CSS tint paints over it.
+    let _: () = msg_send![content, addSubview: frost, positioned: -1isize, relativeTo: std::ptr::null_mut::<AnyObject>()];
+}
+
+/// The `--ease-out` token from tokens.css as a CAMediaTimingFunction.
+/// `functionWithControlPoints::::` has bare colons `msg_send!` can't spell,
+/// hence `send_message`.
+#[cfg(target_os = "macos")]
+unsafe fn ease_out_timing() -> *mut AnyObject {
+    use objc2::runtime::MessageReceiver;
+    use objc2::sel;
+
+    let Some(timing_class) = AnyClass::get(c"CAMediaTimingFunction") else {
+        return std::ptr::null_mut();
+    };
+    (timing_class as *const AnyClass as *mut AnyObject).send_message(
+        sel!(functionWithControlPoints::::),
+        (0.22f32, 1.0f32, 0.36f32, 1.0f32),
+    )
+}
+
+/// The frost NSVisualEffectView installed by `install_frost`, or null.
+#[cfg(target_os = "macos")]
+unsafe fn frost_view(content: *mut AnyObject) -> *mut AnyObject {
+    use objc2::msg_send;
+
+    let Some(effect_class) = AnyClass::get(c"NSVisualEffectView") else {
+        return std::ptr::null_mut();
+    };
+    let subviews: *mut AnyObject = msg_send![content, subviews];
+    if subviews.is_null() {
+        return std::ptr::null_mut();
+    }
+    let count: usize = msg_send![subviews, count];
+    for index in 0..count {
+        let view: *mut AnyObject = msg_send![subviews, objectAtIndex: index];
+        let is_frost: bool = msg_send![view, isKindOfClass: effect_class];
+        if is_frost {
+            return view;
+        }
+    }
+    std::ptr::null_mut()
+}
+
+/// Turn the contentView's layer between flat (0°) and upright (90°). Because
+/// the frost view and the webview are both subviews, one transform carries the
+/// blur, tint, dot, and waveform together — they can't desync or clip. The
+/// frost's length follows the pill's (CSS shrinks the upright pill; the two
+/// animations share the duration and curve).
+#[cfg(target_os = "macos")]
+unsafe fn rotate_content(hud: &WebviewWindow, vertical: bool, animate: bool) {
+    use objc2::msg_send;
+    use objc2_foundation::{NSNumber, NSPoint, NSRect, NSSize, NSString};
+
+    let (window, content) = content_view(hud);
+    if content.is_null() {
+        return;
+    }
+    let layer: *mut AnyObject = msg_send![content, layer];
+    if layer.is_null() {
+        return;
+    }
+
+    // Rotate about the window's center. AppKit anchors view layers at their
+    // corner, so re-center the anchor (re-asserted every turn — layout passes
+    // can reset layer geometry).
+    let frame: NSRect = msg_send![content, frame];
+    let _: () = msg_send![layer, setAnchorPoint: NSPoint::new(0.5, 0.5)];
+    let _: () = msg_send![layer, setPosition: NSPoint::new(
+        frame.origin.x + frame.size.width / 2.0,
+        frame.origin.y + frame.size.height / 2.0,
+    )];
+
+    // CA's +z spins counterclockwise (y-up), so -90° swings the pill's left
+    // end — the record dot — to the top, matching the old CSS rotate(90deg).
+    let angle = if vertical {
+        -std::f64::consts::FRAC_PI_2
+    } else {
+        0.0
+    };
+    let key = NSString::from_str("transform.rotation.z");
+    let target = NSNumber::new_f64(angle);
+
+    if animate {
+        if let Some(animation_class) = AnyClass::get(c"CABasicAnimation") {
+            let animation: *mut AnyObject = msg_send![animation_class, animationWithKeyPath: &*key];
+            if !animation.is_null() {
+                // Start from wherever the layer visibly is right now, so a
+                // turn reversed mid-flight doubles back smoothly.
+                let presentation: *mut AnyObject = msg_send![layer, presentationLayer];
+                let source = if presentation.is_null() {
+                    layer
+                } else {
+                    presentation
+                };
+                let from: *mut AnyObject = msg_send![source, valueForKeyPath: &*key];
+                let _: () = msg_send![animation, setFromValue: from];
+                let _: () = msg_send![animation, setToValue: &*target];
+                let _: () = msg_send![animation, setDuration: TURN_SECS];
+                let timing = ease_out_timing();
+                if !timing.is_null() {
+                    let _: () = msg_send![animation, setTimingFunction: timing];
+                }
+                let _: () = msg_send![layer, addAnimation: animation, forKey: &*key];
+            }
+        }
+    }
+
+    // The upright pill runs shorter (fewer bars); keep the frost hugging it.
+    // Same duration and curve as the turn, on the same Core Animation clock.
+    let frost = frost_view(content);
+    if !frost.is_null() {
+        let length = if vertical {
+            VERTICAL_PILL_LENGTH
+        } else {
+            PILL_SIZE.width
+        };
+        let frost_frame = NSRect::new(
+            NSPoint::new(
+                (WINDOW_SIZE.width - length) / 2.0,
+                (WINDOW_SIZE.height - PILL_SIZE.height) / 2.0,
+            ),
+            NSSize::new(length, PILL_SIZE.height),
+        );
+        if animate {
+            if let Some(context_class) = AnyClass::get(c"NSAnimationContext") {
+                let _: () = msg_send![context_class, beginGrouping];
+                let context: *mut AnyObject = msg_send![context_class, currentContext];
+                if !context.is_null() {
+                    let _: () = msg_send![context, setDuration: TURN_SECS];
+                    let timing = ease_out_timing();
+                    if !timing.is_null() {
+                        let _: () = msg_send![context, setTimingFunction: timing];
+                    }
+                }
+                let animator: *mut AnyObject = msg_send![frost, animator];
+                if !animator.is_null() {
+                    let _: () = msg_send![animator, setFrame: frost_frame];
+                }
+                let _: () = msg_send![context_class, endGrouping];
+            }
+        } else {
+            let _: () = msg_send![frost, setFrame: frost_frame];
+        }
+    }
+
+    // Commit the model value with implicit actions off — the explicit
+    // animation above (when any) owns the visible motion.
+    if let Some(transaction) = AnyClass::get(c"CATransaction") {
+        let _: () = msg_send![transaction, begin];
+        let _: () = msg_send![transaction, setDisableActions: true];
+        let _: () = msg_send![layer, setValue: &*target, forKeyPath: &*key];
+        let _: () = msg_send![transaction, commit];
+    } else {
+        let _: () = msg_send![layer, setValue: &*target, forKeyPath: &*key];
+    }
+    if !animate && !window.is_null() {
+        let _: () = msg_send![window, invalidateShadow];
+    }
+}
+
+/// Ask AppKit to recompute the window shadow from the rendered content — the
+/// shape changed when the pill turned.
+#[cfg(target_os = "macos")]
+unsafe fn invalidate_shadow(hud: &WebviewWindow) {
+    use objc2::msg_send;
+
+    let (window, _) = content_view(hud);
+    if !window.is_null() {
+        let _: () = msg_send![window, invalidateShadow];
+    }
+}
+
 /// macOS: reclass the HUD's NSWindow to a non-activating NSPanel so clicking the
 /// pill or its buttons never steals focus from the meeting app underneath.
 /// Mirrors `dictation.rs::make_hud_nonactivating`.
 #[cfg(target_os = "macos")]
 fn make_nonactivating(hud: &WebviewWindow) {
     use objc2::msg_send;
-    use objc2::runtime::{AnyClass, AnyObject};
 
     let Ok(handle) = hud.ns_window() else {
         return;
